+++ diskover-web/public/js/diskover-dashboard.js
@@ -3,7 +3,7 @@ diskover-web community edition (ce)
 https://github.com/diskoverdata/diskover-community/
 https://diskoverdata.com
 
-Copyright 2017-2021 Diskover Data, Inc.
+Copyright 2017-2022 Diskover Data, Inc.
 "Community" portion of Diskover made available under the Apache 2.0 License found here:
 https://www.diskoverdata.com/apache-license/
  
@@ -118,6 +118,24 @@ $(function () {
             title: {
                 display: true,
                 text: 'Top File Types by Size'
+            },
+            onClick: function(event, clickedElements) {
+                if (clickedElements.length === 0) return;
+                var e = clickedElements[0];
+                var ext = this.data.labels[e._index];
+                //var size = this.data.datasets[0].data[e._index];
+                if (ext == "NULL (no ext)") {
+                    ext = "\"\"";
+                }
+                var pp = encodeURIComponent(escapeHTML(decodeURIComponent(path)));
+                window.location.href = "search.php?submitted=true&p=1&q=extension:" + ext + " AND parent_path:" + pp + "*&doctype=file";
+                return false;
+            },
+            onHover: function (event, legendItem, legend) {
+                $("#topFileTypesBySize-barChart").css("cursor", "pointer");
+            },
+            onLeave: function (event, legendItem, legend) {
+                $("#topFileTypesBySize-barChart").css("cursor", "default");
             }
         }
 
@@ -174,7 +192,7 @@ $(function () {
                     if (ext == "NULL (no ext)") {
                         ext = "\"\"";
                     }
-                    window.location.href = "search.php?submitted=true&p=1&q=extension:" + ext + " AND parent_path:" + encodeURIComponent(escapeHTML(path)) +"*&sort=size&sortorder=desc&sort2=name&sortorder2=asc&doctype=file";
+                    window.location.href = "search.php?submitted=true&p=1&q=extension:" + ext + " AND parent_path:" + encodeURIComponent(escapeHTML(path)) +"*&doctype=file";
                     return false;
                 },
                 onHover: function (event, legendItem, legend) {
@@ -201,6 +219,24 @@ $(function () {
             title: {
                 display: true,
                 text: 'Top File Types by Count'
+            },
+            onClick: function(event, clickedElements) {
+                if (clickedElements.length === 0) return;
+                var e = clickedElements[0];
+                var ext = this.data.labels[e._index];
+                //var size = this.data.datasets[0].data[e._index];
+                if (ext == "NULL (no ext)") {
+                    ext = "\"\"";
+                }
+                var pp = encodeURIComponent(escapeHTML(decodeURIComponent(path)));
+                window.location.href = "search.php?submitted=true&p=1&q=extension:" + ext + " AND parent_path:" + pp + "*&doctype=file";
+                return false;
+            },
+            onHover: function (event, legendItem, legend) {
+                $("#topFileTypesByCount-pieChart").css("cursor", "pointer");
+            },
+            onLeave: function (event, legendItem, legend) {
+                $("#topFileTypesByCount-pieChart").css("cursor", "default");
             }
         }
 
@@ -280,7 +316,7 @@ $(function () {
                     } else if (mtime == "> 2 years") {
                         mtime = "[* TO now/m-2y/d}"
                     }
-                    window.location.href = "search.php?submitted=true&p=1&q=mtime:" + mtime + " AND parent_path:" + encodeURIComponent(escapeHTML(path)) +"*&sort=size&sortorder=desc&sort2=name&sortorder2=asc&doctype=file";
+                    window.location.href = "search.php?submitted=true&p=1&q=mtime:" + mtime + " AND parent_path:" + encodeURIComponent(escapeHTML(path)) +"*&doctype=file";
                     return false;
                 },
                 onHover: function (event, legendItem, legend) {
@@ -324,7 +360,34 @@ $(function () {
                     }
                 }]
             },
-            maintainAspectRatio: false
+            maintainAspectRatio: false,
+            onClick: function(event, clickedElements) {
+                if (clickedElements.length === 0) return;
+                var activeElement = mtimebarChart.getElementAtEvent(event);
+                var mtime = this.data.datasets[activeElement[0]._datasetIndex].label;
+                if (mtime == "0 - 30 days") {
+                    mtime = "[now/m-1M/d TO now/m}"
+                } else if (mtime == "30 - 90 days") {
+                    mtime = "[now/m-3M/d TO now/m-1M/d}"
+                } else if (mtime == "90 - 180 days") {
+                    mtime = "[now/m-6M/d TO now/m-3M/d}"
+                } else if (mtime == "180 days - 1 year") {
+                    mtime = "[now/m-1y/d TO now/m-6M/d}"
+                } else if (mtime == "1 - 2 years") {
+                    mtime = "[now/m-2y/d TO now/m-1y/d}"
+                } else if (mtime == "> 2 years") {
+                    mtime = "[* TO now/m-2y/d}"
+                }
+                var pp = encodeURIComponent(escapeHTML(decodeURIComponent(path)));
+                window.location.href = "search.php?submitted=true&p=1&q=mtime:" + mtime + " AND parent_path:" + pp + "*&doctype=file";
+                return false;
+            },
+            onHover: function (event, legendItem, legend) {
+                $("#mtime-barChart").css("cursor", "pointer");
+            },
+            onLeave: function (event, legendItem, legend) {
+                $("#mtime-barChart").css("cursor", "default");
+            }
         }
 
         var mtimebarChart = new Chart(mtimebarChartCanvas, {